Fixed lots of compiler warnings  Fixed memory leaks

+++ source/guifunctions.cpp
@@ -1577,11 +1577,11 @@ Surface buildGuiIcon( const Surface& image, bool remove = false )
    if ( o.GetPixelFormat().BitsPerPixel() == 32 ) {
       MegaBlitter<4,4,ColorTransform_None, ColorMerger_AlphaOverwrite, SourcePixelSelector_DirectZoom> blitter;
       blitter.setSize( o.w(), o.h(), s.w(), s.h() );
-      blitter.blit( o, s, SPoint((s.w() - blitter.getZoomX() * o.w())/2, (s.h() - blitter.getZoomY() * o.h())/2));
+      blitter.blit( o, s, SPoint(int((s.w() - blitter.getZoomX() * o.w())/2), int((s.h() - blitter.getZoomY() * o.h())/2)));
    } else {
       MegaBlitter<1,4,ColorTransform_None, ColorMerger_AlphaOverwrite, SourcePixelSelector_DirectZoom> blitter;
       blitter.setSize( o.w(), o.h(), s.w(), s.h() );
-      blitter.blit( o, s, SPoint((s.w() - blitter.getZoomX() * o.w())/2, (s.h() - blitter.getZoomY() * o.h())/2));
+      blitter.blit( o, s, SPoint(int((s.w() - blitter.getZoomX() * o.w())/2), int((s.h() - blitter.getZoomY() * o.h())/2)));
    }
 
    if ( remove ) {
